pyproject.toml
[tool.poetry]
name = "nonebot-plugin-cloudsignx"
version = "2.1.2"
description = "Cloudsign_云签到"
authors = ["Monarchdos <admin@ayfre.com>"]
readme = "README.md"
homepage = "https://github.com/Monarchdos/nonebot_plugin_cloudsign"
repository = "https://github.com/Monarchdos/nonebot_plugin_cloudsign"
documentation = "https://github.com/Monarchdos/nonebot_plugin_cloudsign#readme"
keywords = ["nonebot", "nonebot2", "cloudsign"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "*"
nonebot2 = ">=2.0.0-beta.1,<3.0.0"
nonebot-adapter-onebot = ">=2.0.0-beta.1,<3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
